Move YAML writing to separate file
